Flytter overstyrbar utbetaling til egen fil
